Reviews Recap: Move proposals recap in its own adapter
